Adding a new school to JetBrains
Dear Ladies and Gentlemen,

I would like to add the Comprehensive School Gleiberger Land to your list of educational institutions. I am a teacher at the school and also one of three IT officers.
The URL https://ggl-wettenberg.de/ is the URL leading to our learning platform. Our primary website is https://www.gesamtschule-gleiberger-land.de/ . I have attached both files / URLs in the request.
Since we are a public school in Germany and do not show courses as such on our website, it is difficult for me to prove this part of the requirements. However, you can see from the screenshot ( https://imgur.com/a/0PabwB4 ) that the URLs https://www.gesamtschule-gleiberger-land.de/ and https://ggl-wettenberg.de/ are related and confirm what I said earlier.

Yours sincerely
Sebastian Gerla